Add MultidevBranchAwareTrait
