+++ src/Commands/Traits/MultidevBranchAwareTrait.php
@@ -1,0 +1,42 @@
+<?php
+
+namespace Pantheon\TerminusConversionTools\Commands\Traits;
+
+use Pantheon\Terminus\Exceptions\TerminusException;
+
+/**
+ * Trait MultidevBranchAwareTrait.
+ */
+trait MultidevBranchAwareTrait
+{
+    /**
+     * @var string
+     */
+    private string $branch;
+
+    /**
+     * Sets the multidev git branch.
+     *
+     * @param string $branch
+     *
+     * @throws \Pantheon\Terminus\Exceptions\TerminusException
+     */
+    private function setBranch(string $branch): void
+    {
+        if (strlen($branch) > 11) {
+            throw new TerminusException('The git branch name for a multidev env must not exceed 11 characters limit.');
+        }
+
+        $this->branch = $branch;
+    }
+
+    /**
+     * Returns the multidev git branch.
+     *
+     * @return string
+     */
+    private function getBranch(): string
+    {
+        return $this->branch;
+    }
+}
